Improved settings menu + prevent port issue
Settings now can be accessed from within the program so you don't have to delete the settings file.
You can now configure a port the application uses. Additional by default now any free port is used (So no more conflicting ports).
It now warns you if you forgot to enter the path to the save files - many users made this mistake (by simply not filling it out).
Additionally the application won't close anymore if an error is thrown when initialising the GUI - making it easier for users to report the errors.