0.1.7
Added the ability to remove all item drops (in case your game started lagging and you don't know why, it could be because of those drops if you have tons of them).
Save the JSON a bit more compatible (replace NaN with the more correct null).
Corrected the filename of one item.
Removed the content regarding perks which was not working anyways.
Added a loading circle when the user can't interact yet with the data of a specific save.
It is now checked for updates by default.
Closes https://github.com/NetroScript/Graveyard-Keeper-Savefile-Editor/issues/33