Bug Fix when loading item ids with buffers in sub inventories
Closes #73

When loading items in subinventories their ID is extracted as a string.
But in some edge cases for some strings (see capital C) the editor loads this wrongly because the DC4 character is in the string. For those special strings it backups the original binary string buffer it loaded. But now we tried to save an object containing that buffer as a string, which failed.
We now check if the item ID is contained within such a buffer.